+++ LdnServer/Server/LdnServer.cs
@@ -1,11 +1,8 @@
-﻿using LanPlayServer.Stats;
-using LanPlayServer.Stats.Types;
-using NetCoreServer;
+﻿using NetCoreServer;
 using Ryujinx.HLE.HOS.Services.Ldn.Types;
 using System;
 using System.Collections.Concurrent;
 using System.Collections.Generic;
-using System.Linq;
 using System.Net;
 using System.Net.Sockets;
 using System.Threading;
@@ -13,8 +10,9 @@ using System.Threading.Tasks;
 
 namespace LanPlayServer
 {
-    public class LdnServer : TcpServer
+    public partial class LdnServer : TcpServer
     {
+        public const long OneHour = 3600000;
         public const int InactivityPingFrequency = 10000;
 
         private readonly ConcurrentDictionary<string, HostedGame> _hostedGames = new();
@@ -31,69 +29,17 @@ namespace LanPlayServer
             Task.Run(BackgroundDumpTask);
         }
 
-        public HostedGame CreateGame(string id, NetworkInfo info, AddressList dhcpConfig, string oldOwnerID)
-        {
-            id = id.ToLower();
-            HostedGame game = new(id, info, dhcpConfig);
-            bool idTaken = false;
-
-            _hostedGames.AddOrUpdate(id, game, (id, oldGame) =>
-            {
-                if (oldGame.OwnerId == oldOwnerID)
-                {
-                    oldGame.Close();
-
-                    Statistics.RemoveGameAnalytics(oldGame);
-
-                    return game;
-                }
-                else
-                {
-                    game.Close();
-                    idTaken = true;
-
-                    Console.WriteLine($"id Taken: {id}");
-                    return oldGame;
-                }
-            });
-
-            if (idTaken)
-            {
-                return null;
-            }
-
-            Statistics.AddGameAnalytics(game);
-
-            return game;
-        }
-
-        public HostedGame FindGame(string id)
-        {
-            id = id.ToLower();
-
-            _hostedGames.TryGetValue(id, out HostedGame result);
-
-            return result;
-        }
-
-        public HostedGame[] All()
-        {
-            return _hostedGames.Values.ToArray();
-        }
-
         public int Scan(ref NetworkInfo[] info, ScanFilter filter, string passphrase, HostedGame exclude)
         {
             KeyValuePair<string, HostedGame>[] all = _hostedGames.ToArray();
 
             int results = 0;
-
-            int gameCount = all.Length;
             int playerCount = 0;
 
             long currentTime = DateTimeOffset.UtcNow.ToUnixTimeMilliseconds();
-            long oneHour = 3600000;
+
             // Games older than this are probably bugged "ghost" lobbies, still rarely happens, there's probably still a lock issue somewhere
-            long minTime = currentTime - (oneHour * 16);
+            long minTime = currentTime - (OneHour * 16);
 
             for (int i = 0; i < all.Length; i++)
             {
@@ -183,69 +129,16 @@ namespace LanPlayServer
             return results;
         }
 
-        public void CloseGame(string id)
-        {
-            _hostedGames.Remove(id.ToLower(), out HostedGame removed);
-            removed?.Close();
+        protected override TcpSession CreateSession() 
+            => new LdnSession(this);
 
-            if (removed != null)
-            {
-                Console.WriteLine($"Removing from analytics: {id}");
-                Statistics.RemoveGameAnalytics(removed);
-            }
-        }
-
-        protected override TcpSession CreateSession()
-        {
-            return new LdnSession(this);
-        }
-
-        protected override void OnError(SocketError error)
-        {
-            Console.WriteLine($"LDN TCP server caught an error with code {error}");
-        }
+        protected override void OnError(SocketError error) => Console.WriteLine($"LDN TCP server caught an error with code {error}");
 
         public override bool Stop()
         {
             _cancel.Cancel();
 
             return base.Stop();
-        }
-
-        private async Task BackgroundPingTask()
-        {
-            while (!IsDisposed)
-            {
-                foreach (KeyValuePair<Guid, TcpSession> session in Sessions)
-                {
-                    (session.Value as LdnSession).Ping();
-                }
-
-                try
-                {
-                    await Task.Delay(InactivityPingFrequency, _cancel.Token);
-                }
-                catch (TaskCanceledException)
-                {
-                    return;
-                }
-            }
-        }
-
-        private async Task BackgroundDumpTask()
-        {
-            while (!IsDisposed)
-            {
-                await Task.Delay(5000, _cancel.Token);
-                try
-                {
-                    await StatsDumper.DumpAll(_hostedGames);
-                }
-                catch (Exception e)
-                {
-                    Console.WriteLine(e);
-                }
-            }
         }
     }
 }
