+++ LdnServer/Session/Handlers/Disconnect.cs
@@ -1,0 +1,9 @@
+﻿using LanPlayServer.Network.Types;
+
+namespace LanPlayServer
+{
+    public partial class LdnSession
+    {
+        private void HandleDisconnect(LdnHeader header, DisconnectMessage message) => DisconnectFromGame();
+    }
+}
